Refactor navigation. Migrate to Cicerone. Migrate to splashScreen dep.

+++ data/network/src/main/kotlin/com/grappim/network/model/waybill/WaybillProductsRequestDTO.kt
@@ -10,5 +10,5 @@ data class WaybillProductsRequestDTO(
   @SerialName("offset")
   val offset: Int,
   @SerialName("waybillId")
-  val waybillId: Int
+  val waybillId: Long
 )
